LayoutTests/fetch/header-constructor-overriden-expected.txt

This test should create a Headers, override the Headers property and try creating new Headers from it.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS headers.get('hello') is "world"
PASS newHeaders.get('hello') is "world"
PASS response.headers.get('hello') is "world"
PASS successfullyParsed is true

TEST COMPLETE
